% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldet}
\alias{ldet}
\title{Log determinant of a positive definite matrix}
\usage{
ldet(X)
}
\arguments{
\item{X}{Square positive definite matrix}
}
\value{
log|X|
}
\description{
Uses Armadillo's log_det for numerical stability.
}
