% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lgammap_export}
\alias{lgammap_export}
\title{Log of the multivariate Gamma function}
\usage{
lgammap_export(a, p)
}
\arguments{
\item{a}{Argument, must be > (p-1)/2}

\item{p}{Dimension}
}
\value{
log Gamma_p(a)
}
\description{
Defined as log Gamma_p(a) = p*(p-1)/4 * log(pi) plus the sum over
i = 1, ..., p of log Gamma(a - (i-1)/2). Required for the normalizing
constant of the Wishart and inverse-Wishart distributions.
}
