% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode_alphaEM_gamma}
\alias{mode_alphaEM_gamma}
\title{EM algorithm to find the posterior mode of alpha, gamma (p = 1) case}
\usage{
mode_alphaEM_gamma(
  n,
  xbar,
  ldetxbarg,
  beta = 0,
  eta = 1,
  kappa = 0,
  mu0 = 1,
  tol = 1e-06,
  prnt = FALSE,
  max_em_iter = 1000L
)
}
\arguments{
\item{n}{Number of scalar observations}

\item{xbar}{Sample mean of the observations}

\item{ldetxbarg}{mean(log(x))}

\item{beta}{Gamma prior shape, must be >= 0 (0 = improper)}

\item{eta}{Gamma prior rate, must be >= 0}

\item{kappa}{Prior strength on mu, must be >= 0 (0 = improper on mu)}

\item{mu0}{Prior center for mu, must be > 0}

\item{tol}{Convergence tolerance}

\item{prnt}{If TRUE, print EM iterates and bounds}

\item{max_em_iter}{Maximum number of EM iterations}
}
\value{
A numeric vector \code{c(ahat, log f*(ahat))}
}
\description{
Scalar (p = 1) analogue of \code{mode_alphaEM()}. See the C++ doc
comment for details on the bisection-style ascent used here and the
n = 2, kappa = 0, beta = 0 structural boundary case.
}
