% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwishart}
\alias{rwishart}
\title{Generate draws from a Wishart distribution}
\usage{
rwishart(n, p, nu, Sigma)
}
\arguments{
\item{n}{Number of draws}

\item{p}{Dimension of the matrix, must be >= 2}

\item{nu}{Degrees of freedom, must be > p - 1}

\item{Sigma}{Scale matrix (p x p), must be positive definite}
}
\value{
A cube of size (p, p, n) containing the Wishart draws
}
\description{
Uses the Bartlett decomposition: if L is the lower Cholesky factor of
Sigma, then A = L * T satisfies W = A * A^T ~ Wishart_p(nu, Sigma),
where T is lower triangular with T(i,i) ~ sqrt(Chi^2(nu - i)) and
T(i,j) ~ N(0,1) for i > j.
}
