% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishartinference-package.R
\docType{package}
\name{wishartinference-package}
\alias{wishartinference}
\alias{wishartinference-package}
\title{wishartinference: Bayesian Inference for the Wishart Distribution Shape Parameter}
\description{
Posterior inference for the shape parameter alpha and mean matrix mu
in the model X_i ~ Wishart_p(2*alpha, Sigma).
}
\author{
\strong{Maintainer}: Hanqi Shi \email{hshi1@swarthmore.edu}

Authors:
\itemize{
  \item Hanqi Shi \email{hshi1@swarthmore.edu}
  \item Phil Everson \email{peverso1@swarthmore.edu}
}

}
\keyword{internal}
